src/agenticaiorchestrator/requirement.txt
crewai>=0.1.0
PyPDF2>=3.0.0
PyYAML>=6.0
requests>=2.28.0
pydantic>=2.0.0
langchain>=0.0.300
langchain-community>=0.0.1